# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.FileUploadResponse:
    properties:
      file:
        $ref: '#/definitions/service.FileResponse'
      id:
        type: integer
    type: object
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.AnniversaryCreateRequest:
    properties:
      calendar:
        enum:
        - solar
        - lunar
        type: string
      date:
        description: YYYY-MM-DD格式
        type: string
      description:
        type: string
      title:
        type: string
    required:
    - calendar
    - date
    - title
    type: object
  service.AnniversaryListResponse:
    properties:
      anniversaries:
        items:
          $ref: '#/definitions/service.FrontendAnniversary'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
      totalCount:
        type: integer
      totalPages:
        type: integer
    type: object
  service.AnniversaryUpdateRequest:
    properties:
      calendar:
        enum:
        - solar
        - lunar
        type: string
      date:
        description: YYYY-MM-DD格式
        type: string
      description:
        type: string
      title:
        type: string
    type: object
  service.FileResponse:
    properties:
      id:
        type: integer
      mime_type:
        type: string
      name:
        type: string
      size:
        type: integer
      thumbnail:
        type: string
      url:
        type: string
    type: object
  service.FrontendAnniversary:
    properties:
      calendar:
        description: solar/lunar
        type: string
      date:
        description: MM-DD
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /anniversaries:
    get:
      description: 分页获取纪念日列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.AnniversaryListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取纪念日列表
      tags:
      - anniversaries
    post:
      consumes:
      - application/json
      description: 创建一个新的纪念日
      parameters:
      - description: 纪念日信息
        in: body
        name: anniversary
        required: true
        schema:
          $ref: '#/definitions/service.AnniversaryCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.FrontendAnniversary'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: 创建纪念日
      tags:
      - anniversaries
  /anniversaries/{id}:
    delete:
      description: 删除指定ID的纪念日
      parameters:
      - description: 纪念日ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: 删除纪念日
      tags:
      - anniversaries
    put:
      consumes:
      - application/json
      description: 更新指定ID的纪念日信息
      parameters:
      - description: 纪念日ID
        in: path
        name: id
        required: true
        type: integer
      - description: 纪念日更新信息
        in: body
        name: anniversary
        required: true
        schema:
          $ref: '#/definitions/service.AnniversaryUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.FrontendAnniversary'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - OAuth2Password: []
      summary: 更新纪念日
      tags:
      - anniversaries
...
  /file/{id}:
    get:
      description: Returns the file content as a stream. The browser will either preview
        or download based on Content-Disposition.
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File stream
          schema:
            type: file
        "400":
          description: Invalid file ID format
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Failed to read file from storage
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Get a file by ID
      tags:
      - files
  /file/upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a file to the server. Supports optional fields like filename,
        path, mimeType, hash, and thumbnail dimensions.
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Storage path prefix
        in: formData
        name: path
        type: string
      - description: File content hash (e.g., SHA256)
        in: formData
        name: hash
        type: string
      - description: Desired thumbnail width (for image processing)
        in: formData
        name: thumbnailWidth
        type: string
      - description: Desired thumbnail height (for image processing)
        in: formData
        name: thumbnailHeight
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: File uploaded successfully
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.FileUploadResponse'
              type: object
        "400":
          description: Missing or invalid file
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal server error during file saving or URL generation
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Upload a file
      tags:
      - files
...
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      admin: Grants read and write access to administrative information
      read: Grants read access
      write: Grants write access
    tokenUrl: /api/v1/user/token
    type: oauth2
swagger: "2.0"